requirements.txt
openai
pyaudio
gTTS
assemblyai[extras]
websocket-client
streamlit
python-dotenv
openai
